requirements.txt
Flask==2.3.2
gunicorn==21.2.0
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.3
Werkzeug==3.0.3
openai==1.58.1
